+++ joulescope_ui/plugins/test/plugins/__init__.py
@@ -1,0 +1,1 @@
+# Joulescope UI plugin directory.  Add plugin packages here